Lab05/img-hidden/requirements.txt
Pillow
cryptogarphy